pyproject.toml
[project]
name = "change-data-capture"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Daniel-Lars", email = "daniel2wendel@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "kafka-python>=2.2.15",
    "requests>=2.32.5",
]

[project.scripts]
change-data-capture = "change_data_capture:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
